Renamed vrpc_local.py to VrpcLocal.py
