+++ .travis.yml
@@ -18,7 +18,7 @@ matrix:
       install:
         - BUILD_TEST=1 pip install .
       script:
-        - python test/python/vrpc_local_test.py
+        - python test/python/VrpcLocal_test.py
 
 notifications:
   email:
